+++ .gitignore
@@ -1,0 +1,2 @@
+
+android/Sara/common/src/main/java/com/example/common/API_KEY.kt
