+++ android/Sara/settings.gradle
@@ -14,3 +14,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "Sara"
 include ':app'
+include ':common'
